requirements.txt
dash>=2.16
dash-bootstrap-components>=1.6
flask>=3.0
flask-login>=0.6
flask-caching>=2.1
pandas>=2.2
numpy>=1.26
requests>=2.32
python-dotenv>=1.0

plotly>=5.18
kaleido>=0.2.1
reportlab>=4.0
